refactor: 통합테스트 추상 클래스로 분리
